file name change

+++ Spira Unleashed/Linux/procinfo.h
@@ -15,7 +15,7 @@ public:
     ProcInfo();
 
     pid_t getProcessHandle() const;
-    long getModuleBaseAddress() const;
+    uintptr_t getModuleBaseAddress() const;
 
 private:
     void getProcId();
@@ -27,7 +27,7 @@ private:
 
     pid_t procId = -1;
     pid_t pHandle = -1;
-    long modBaseAddr = 0;
+    uintptr_t modBaseAddr = 0;
 };
 
 #endif // PROCINFO_H
